Cleanup:
- Updated README files for parent project, Client, and Server.
- Updated folder structure
