+++ data/go.mod
@@ -1,0 +1,5 @@
+module github.com/matthi01/ledn_dashboard/accounts
+
+go 1.16
+
+require go.mongodb.org/mongo-driver v1.7.0 // indirect
